*** empty log message ***

+++ magick/compress.c
@@ -1017,8 +1017,8 @@ MagickExport unsigned int Huffman2DEncodeImage(ImageInfo *image_info,
 %
 %  The format of the LZWEncodeImage method is:
 %
-%      unsigned int LZWEncodeImage(Image *image,
-%        const unsigned int number_pixels,unsigned char *pixels)
+%      unsigned int LZWEncodeImage(Image *image,const size_t number_pixels,
+%        unsigned char *pixels)
 %
 %  A description of each parameter follows:
 %
@@ -1027,8 +1027,7 @@ MagickExport unsigned int Huffman2DEncodeImage(ImageInfo *image_info,
 %
 %    o image: The image.
 %
-%    o number_pixels:  An unsigned interger that specifies the number of
-%      pixels to compress.
+%    o number_pixels:  A value that specifies the number of pixels to compress.
 %
 %    o pixels: The address of an unsigned array of characters containing the
 %      pixels to compress.
@@ -1036,7 +1035,7 @@ MagickExport unsigned int Huffman2DEncodeImage(ImageInfo *image_info,
 %
 */
 MagickExport unsigned int LZWEncodeImage(Image *image,
-  const unsigned int number_pixels,unsigned char *pixels)
+  const size_t number_pixels,unsigned char *pixels)
 {
 #define LZWClr  256  /* Clear Table Marker */
 #define LZWEod  257  /* End of Data marker */
@@ -1104,7 +1103,7 @@ MagickExport unsigned int LZWEncodeImage(Image *image,
   next_index=LZWEod+1;
   code_width=9;
   last_code=pixels[0];
-  for (i=1; i < (int) number_pixels; i++)
+  for (i=1; i < number_pixels; i++)
   {
     /*
       Find string.
@@ -1168,7 +1167,7 @@ MagickExport unsigned int LZWEncodeImage(Image *image,
 }
 #else
 MagickExport unsigned int LZWEncodeImage(Image *image,
-  const unsigned int number_pixels,unsigned char *pixels)
+  const size_t number_pixels,unsigned char *pixels)
 {
   assert(image != (Image *) NULL);
   assert(image->signature == MagickSignature);
@@ -1194,8 +1193,8 @@ MagickExport unsigned int LZWEncodeImage(Image *image,
 %
 %  The format of the PackbitsEncodeImage method is:
 %
-%      unsigned int PackbitsEncodeImage(Image *image,
-%        const unsigned int number_pixels,unsigned char *pixels)
+%      unsigned int PackbitsEncodeImage(Image *image,const size_t number_pixels,
+%        unsigned char *pixels)
 %
 %  A description of each parameter follows:
 %
@@ -1204,8 +1203,7 @@ MagickExport unsigned int LZWEncodeImage(Image *image,
 %
 %    o image: The image.
 %
-%    o number_pixels:  An unsigned integer that specifies the number of
-%      pixels to compress.
+%    o number_pixels:  A value that specifies the number of pixels to compress.
 %
 %    o pixels: The address of an unsigned array of characters containing the
 %      pixels to compress.
@@ -1213,7 +1211,7 @@ MagickExport unsigned int LZWEncodeImage(Image *image,
 %
 */
 MagickExport unsigned int PackbitsEncodeImage(Image *image,
-  const unsigned int number_pixels,unsigned char *pixels)
+  const size_t number_pixels,unsigned char *pixels)
 {
   register int
     count,
@@ -1332,9 +1330,8 @@ MagickExport unsigned int PackbitsEncodeImage(Image *image,
 %
 %  The format of the ZLIBEncodeImage method is:
 %
-%      unsigned int ZLIBEncodeImage(Image *image,
-%        const unsigned long number_pixels,const unsigned int quality,
-%        unsigned char *pixels)
+%      unsigned int ZLIBEncodeImage(Image *image,const size_t number_pixels,
+%        const unsigned int quality,unsigned char *pixels)
 %
 %  A description of each parameter follows:
 %
@@ -1344,8 +1341,7 @@ MagickExport unsigned int PackbitsEncodeImage(Image *image,
 %    o file: The address of a structure of type FILE.  ZLIB encoded pixels
 %      are written to this file.
 %
-%    o number_pixels:  An unsigned integer that specifies the number of
-%      pixels to compress.
+%    o number_pixels:  A value that specifies the number of pixels to compress.
 %
 %    o quality: the compression level (0-100).
 %
@@ -1355,8 +1351,7 @@ MagickExport unsigned int PackbitsEncodeImage(Image *image,
 %
 */
 MagickExport unsigned int ZLIBEncodeImage(Image *image,
-  const unsigned long number_pixels,const unsigned int quality,
-  unsigned char *pixels)
+  const size_t number_pixels,const unsigned int quality,unsigned char *pixels)
 {
   int
     status;
@@ -1408,8 +1403,7 @@ MagickExport unsigned int ZLIBEncodeImage(Image *image,
 }
 #else
 MagickExport unsigned int ZLIBEncodeImage(Image *image,
-  const unsigned long number_pixels,const unsigned int quality,
-  unsigned char *pixels)
+  const size_t number_pixels,const unsigned int quality,unsigned char *pixels)
 {
   assert(image != (Image *) NULL);
   assert(image->signature == MagickSignature);
